carLogoDetection/requirements.txt
Django
djangorestframework
Pillow
torch
torchvision
uwsgi
